refactor: rename switch
