+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-11-16T13:48:39.789Z\n"
-"PO-Revision-Date: 2023-11-16T13:48:39.789Z\n"
+"POT-Creation-Date: 2023-11-16T14:05:43.332Z\n"
+"PO-Revision-Date: 2023-11-16T14:05:43.332Z\n"
 
 msgid "Something went wrong"
 msgstr "Something went wrong"
@@ -259,8 +259,11 @@ msgstr ""
 "Something went wrong whilst loading the requested job. Make sure it has not "
 "been deleted and try refreshing the page."
 
-msgid "Toggle job"
-msgstr "Toggle job"
+msgid "Disable"
+msgstr "Disable"
+
+msgid "Enable"
+msgstr "Enable"
 
 msgid "New Job"
 msgstr "New Job"
